# Conflicts:
#	src/Main.java
#	src/SalesManager.java

+++ src/SalesManager.java
@@ -1,13 +1,12 @@
 public class SalesManager {
-    protected int[] sales;
-
-    public SalesManager(int[] sales) {
+    protected long[] sales;
+    public SalesManager(long[] sales) {
         this.sales = sales;
     }
 
-    public int max() {
-        int max = -1;
-        for (int sale : sales) {
+    public long max() {
+        long max = -1;
+        for (long sale : sales) {
             if (sale > max) {
                 max = sale;
             }
@@ -15,11 +14,11 @@ public class SalesManager {
         return max;
     }
 
-    public int averageSales() {
-        int max = sales[0];
-        int min = sales[0];
-        int sum = 0;
-        for (int sale : sales) {
+    public long averageSales() {
+        long max = sales[0];
+        long min = sales[0];
+        long sum = 0;
+        for (long sale : sales) {
             sum = sum + sale;
             if (sale > max) {
                 max = sale;
